Reorganize code structure into separate modules
- Split operations into dedicated modules by type.
- Removed operations.py.
